+++ tauri-app/src/lib/queries/keys.ts
@@ -1,3 +1,4 @@
 export const QKEY_VAULTS = 'vaults';
 export const QKEY_VAULT = 'vault';
 export const QKEY_VAULT_CHANGES = 'vaultBalanceChanges';
+export const QKEY_ORDERS = 'orders';
