+++ tauri-app/src/lib/queries/vaultList.ts
@@ -1,5 +1,4 @@
 import type { Vault } from '$lib/typeshare/vaultsList';
-// import type { Vault as VaultDetail } from '$lib/typeshare/vaultDetail';
 import { invoke } from '@tauri-apps/api';
 import { DEFAULT_PAGE_SIZE } from './constants';
 import { mockIPC } from '@tauri-apps/api/mocks';
